allow fasta/ref mismatches when normalising external VCFs
External VCFs such as cosmic and dbsnp sometimes contain variants in the chrY-PAR region. To improve mapping, our reference contains only the X-PAR, and the Y-PAR is masked with `N`s. Therefore, the `ALT` field in the external resources can be the actual nucleotides, whereas they are `N` in our reference fasta. Setting `-n` in `vt normalize` allows for such mismatches.